refactor: convert to marker interface
