+++ packages/backend/src/repositories/staff/StaffRepository.ts
@@ -1,14 +1,6 @@
 import { Staff } from '@/entities/Staff';
-import { ReadOptions } from '@/repositories/types/ReadOptions';
+import { Repository } from '@/repositories/Repository';
 
-export interface StaffRepository {
-
-  create(staff: Staff): Promise<Staff>;
-
-  read(staff: Staff, readOptions?: ReadOptions): Promise<Staff[]>;
-
-  update(source: Staff, destination: Staff): Promise<boolean>;
-
-  delete(staff: Staff): Promise<boolean>;
+export interface StaffRepository extends Repository<Staff> {
 
 }
